Ensure locs_x and locs_y are not modified in-place
In Julia 0.5, the behavior of the plot function is changed in that the existing code modified the values of the locs_x and locs_y vectors in-place when performing operations such as element-wise addition and element-wise subtraction on those vectors.  In Julia 0.4, these element-wise operations created copies of the array on their own.  This commit makes explicit copies of these two vectors within the plot function prior to performing operations that now modify arrays in-place, and hence preserving the user-facing behavior of the plot API between 0.4 and 0.5.