Use LightGraphs instead of Graphs in README.md
This readme still used mostly Graphs.jl instead of LightGraphs.jl. As
the former package is deprecated and GraphPlots.jl does not work with it
anymore, every occurrence of Graphs has changed to LightGraphs.

Some functions functions such as num_vertices or out_degree had to be
changed for that.